frontend/requirements.txt
flask==2.2.3
python-dotenv==0.15.0
flask-login==0.6.2